requirements.txt
pydantic
pydantic-settings
python-dotenv
pandas
openpyxl
xlrd>=2.0.1
python-docx
pillow
rapidfuzz
streamlit
openai
pytesseract
tenacity
httpx
fastapi
uvicorn
python-multipart
PyYAML
pytest>=7.0.0
pytest-cov
